+++ src/Lpf2/Device.cpp
@@ -20,6 +20,9 @@
 #include "Lpf2/Devices/BasicMotor.hpp"
 #include "Lpf2/Devices/ColorSensor.hpp"
 #include "Lpf2/Devices/DistanceSensor.hpp"
+#include "Lpf2/Devices/HubLED.hpp"
+#include "Lpf2/Devices/Accelerometer.hpp"
+#include "Lpf2/Devices/Gyroscope.hpp"
 
 namespace Lpf2
 {
@@ -32,5 +35,8 @@ namespace Lpf2
         // TODO: Add colordistanceSensor here! (above Color and Distance sensors)
         TechnicColorSensor::registerFactory(DeviceRegistry::instance());
         TechnicDistanceSensor::registerFactory(DeviceRegistry::instance());
+        HubLED::registerFactory(DeviceRegistry::instance());
+        HubAccelerometer::registerFactory(DeviceRegistry::instance());
+        HubGyroscope::registerFactory(DeviceRegistry::instance());
     }
 };