+++ src/Lpf2/Devices/ColorSensor.cpp
@@ -57,4 +57,41 @@ namespace Lpf2::Devices
     {
         return ColorIDX((int)m_port.getValue(0, 0));
     }
+
+    float TechnicColorSensor::getReflectivity()
+    {
+        return m_port.getValue(1, 0);
+    }
+
+    void TechnicColorSensor::getRGB(uint16_t &r, uint16_t &g, uint16_t &b, uint16_t &i)
+    {
+        r = (uint16_t)m_port.getValue(5, 0);
+        g = (uint16_t)m_port.getValue(5, 1);
+        b = (uint16_t)m_port.getValue(5, 2);
+        i = (uint16_t)m_port.getValue(5, 3);
+    }
+
+    void TechnicColorSensor::getHSV(uint16_t &h, uint16_t &s, uint16_t &v)
+    {
+        h = (uint16_t)m_port.getValue(6, 0);
+        s = (uint16_t)m_port.getValue(6, 1);
+        v = (uint16_t)m_port.getValue(6, 2);
+    }
+
+    void TechnicColorSensor::setLight(uint8_t l1, uint8_t l2, uint8_t l3)
+    {
+        if (l1 > 100)
+            l1 = 100;
+        if (l2 > 100)
+            l2 = 100;
+        if (l3 > 100)
+            l3 = 100;
+
+        std::vector<uint8_t> data;
+        data.push_back(l1);
+        data.push_back(l2);
+        data.push_back(l3);
+
+        m_port.writeData(LIGHT_MODE, data);
+    }
 }; // namespace Lpf2::Devices